pitron practice

+++ week-2/module 6.5/practice day 1/Ali baba & puzzels.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main()
 {
-   long long int a , b , c ,d;
+    int a , b , c ,d;
     scanf("%d %d %d %d",&a ,&b,&c,&d);
     if (a+b*c==d)
     {
